Методы для зудожника перенесены в компонент Painter в папке API
