+++ react-web-client/src/redux/API/PainterStyles.jsx
