+++ react-web-client/src/redux/state.js
@@ -1,6 +1,6 @@
-import { rerenderEntireTree } from "../render";
 import * as axios from "axios";
-import { getPainterStyle } from "./API/PainterStylesMethods";
+import { getPainterStyle } from "./API/PainterStyles";
+import { getPainterByCount } from "./API/Painter";
 
 let state = {
     contentPage: {
@@ -38,65 +38,7 @@ export let initializeData = () => {
     // rerenderEntireTree(state);
 }
 
-// TODO: вынести в отдельный компонент
-export let setPainter = (name, age, description, styleName) => {
-    let styleId = state.contentPage.painterStyles.find(ps => ps.name === styleName).id
 
-    let painter = {
-        id: newGuid(),
-        name,
-        age,
-        description,
-        styleId: styleId,
-        booksIds: [
-            newGuid()
-        ]
-    }
-    axios.post('https://localhost:44394/api/Painter', painter).then(response => {
-        console.log('add painter', response);
-    })
-
-    getPainterByCount(15);
-    // rerenderEntireTree(state);
-}
-
-export let getPainterByCount = (count) => {
-    axios.get('https://localhost:44394/api/Painter/GetPainters/take/' + count + '/skip/0').then(response => {
-        console.log('get painter', response);
-        state.contentPage.painters = response.data.previewPainters;
-    })
-    return state.contentPage.painterStyles;
-}
-
-export let deletePainter = (name) => {
-    let painterId = state.contentPage.painters.find(ps => ps.name === name).id
-
-    axios.delete('https://localhost:44394/api/Painter/' + painterId).then(response => {
-        console.log('delete painter', response);
-    })
-    initializeData();
-}
-
-export let updatePainter = (oldName, name, age, description, styleName) => {
-    let oldPainter = state.contentPage.painters.find(ps => ps.name === oldName)
-    let styleId = state.contentPage.painterStyles.find(ps => ps.name === styleName).id
-
-    let painter = {
-        id: oldPainter.id,
-        name,
-        age,
-        description,
-        styleId: styleId,
-        booksIds: [
-            newGuid()
-        ]
-    }
-    axios.put('https://localhost:44394/api/Painter/', painter).then(response => {
-        console.log('update painter', response);
-    })
-
-    initializeData();
-}
 
 // TODO: вынести в отдельный компонент
 export let addPublisher = (name) => {
